chore(notifications): move to /packages
