+++ packages/components/src/xm-notifications/notifications/notifications.component.scss
